Split RPC schemas into multiple files (#1385)
* schemas moved to trace_api.py

* schemas moved to transactions.py

* add schemas/rpc/block.py

* delete duplicates

* file rename

* rename ContractClass -> DeprecatedContractClass

* event schema moved to event.py

* everything moved from gateway.py

* rm gateway.py

* styling

* StarknetBlockWithTxHashesSchema moved to block.py

* lint

* change List field type in SierraContractClassSchema

* add breaking changes in migration guide

* circular import fix, general.py added
