+++ starknet_py/net/client_models.py
@@ -762,7 +762,7 @@ class EntryPointsByType:
 
 
 @dataclass
-class ContractClass:
+class DeprecatedContractClass:
     """
     Dataclass representing contract declared to Starknet.
     """
@@ -773,7 +773,7 @@ class ContractClass:
 
 
 @dataclass
-class CompiledContract(ContractClass):
+class DeprecatedCompiledContract(DeprecatedContractClass):
     """
     Dataclass representing ContractClass with required abi.
     """
@@ -812,7 +812,7 @@ class SierraContractClass:
     """
 
     contract_class_version: str
-    sierra_program: List[str]
+    sierra_program: List[int]
     entry_points_by_type: SierraEntryPointsByType
     abi: Optional[str] = None
