+++ starknet_py/net/full_node_client.py
@@ -12,9 +12,9 @@ from starknet_py.net.client_models import (
     BlockStateUpdate,
     BlockTransactionTrace,
     Call,
-    ContractClass,
     DeclareTransactionResponse,
     DeployAccountTransactionResponse,
+    DeprecatedContractClass,
     EstimatedFee,
     EventsChunk,
     Hash,
@@ -51,29 +51,35 @@ from starknet_py.net.models.transaction import (
     DeployAccount,
     Invoke,
 )
-from starknet_py.net.schemas.rpc import (
+from starknet_py.net.schemas.rpc.block import (
     BlockHashAndNumberSchema,
     BlockStateUpdateSchema,
-    BlockTransactionTraceSchema,
-    ContractClassSchema,
-    DeclareTransactionResponseSchema,
-    DeployAccountTransactionResponseSchema,
-    EstimatedFeeSchema,
-    EventsChunkSchema,
     PendingBlockStateUpdateSchema,
     PendingStarknetBlockSchema,
     PendingStarknetBlockWithReceiptsSchema,
     PendingStarknetBlockWithTxHashesSchema,
-    SentTransactionSchema,
-    SierraContractClassSchema,
-    SimulatedTransactionSchema,
     StarknetBlockSchema,
     StarknetBlockWithReceiptsSchema,
     StarknetBlockWithTxHashesSchema,
+)
+from starknet_py.net.schemas.rpc.contract import (
+    DeprecatedContractClassSchema,
+    SierraContractClassSchema,
     SyncStatusSchema,
+)
+from starknet_py.net.schemas.rpc.event import EventsChunkSchema
+from starknet_py.net.schemas.rpc.general import EstimatedFeeSchema
+from starknet_py.net.schemas.rpc.trace_api import (
+    BlockTransactionTraceSchema,
+    SimulatedTransactionSchema,
+    TransactionTraceSchema,
+)
+from starknet_py.net.schemas.rpc.transactions import (
+    DeclareTransactionResponseSchema,
+    DeployAccountTransactionResponseSchema,
+    SentTransactionSchema,
     TransactionReceiptSchema,
     TransactionStatusResponseSchema,
-    TransactionTraceSchema,
     TypesOfTransactionsSchema,
 )
 from starknet_py.transaction_errors import TransactionNotReceivedError
@@ -548,7 +554,7 @@ class FullNodeClient(Client):
         class_hash: Hash,
         block_hash: Optional[Union[Hash, Tag]] = None,
         block_number: Optional[Union[int, Tag]] = None,
-    ) -> Union[SierraContractClass, ContractClass]:
+    ) -> Union[SierraContractClass, DeprecatedContractClass]:
         block_identifier = get_block_identifier(
             block_hash=block_hash, block_number=block_number
         )
@@ -566,7 +572,7 @@ class FullNodeClient(Client):
                 SierraContractClass,
                 SierraContractClassSchema().load(res),
             )
-        return cast(ContractClass, ContractClassSchema().load(res))
+        return cast(DeprecatedContractClass, DeprecatedContractClassSchema().load(res))
 
     # Only RPC methods
 
@@ -625,7 +631,7 @@ class FullNodeClient(Client):
         contract_address: Hash,
         block_hash: Optional[Union[Hash, Tag]] = None,
         block_number: Optional[Union[int, Tag]] = None,
-    ) -> Union[SierraContractClass, ContractClass]:
+    ) -> Union[SierraContractClass, DeprecatedContractClass]:
         """
         Get the contract class definition in the given block at the given address
 
@@ -651,7 +657,7 @@ class FullNodeClient(Client):
                 SierraContractClass,
                 SierraContractClassSchema().load(res),
             )
-        return cast(ContractClass, ContractClassSchema().load(res))
+        return cast(DeprecatedContractClass, DeprecatedContractClassSchema().load(res))
 
     async def get_contract_nonce(
         self,
