+++ starknet_py/proxy/contract_abi_resolver.py
@@ -12,7 +12,7 @@ from starknet_py.constants import (
 )
 from starknet_py.net.client import Client
 from starknet_py.net.client_errors import ClientError, ContractNotFoundError
-from starknet_py.net.client_models import ContractClass, SierraContractClass
+from starknet_py.net.client_models import DeprecatedContractClass, SierraContractClass
 from starknet_py.net.models import Address
 from starknet_py.proxy.proxy_check import (
     ArgentProxyCheck,
@@ -141,17 +141,17 @@ class ContractAbiResolver:
 
     @staticmethod
     def _get_cairo_version(
-        contract_class: Union[ContractClass, SierraContractClass]
+        contract_class: Union[DeprecatedContractClass, SierraContractClass]
     ) -> int:
         return 1 if isinstance(contract_class, SierraContractClass) else 0
 
     @staticmethod
     def _get_abi_from_contract_class(
-        contract_class: Union[ContractClass, SierraContractClass]
+        contract_class: Union[DeprecatedContractClass, SierraContractClass]
     ) -> AbiDictList:
         return (
             cast(AbiDictList, contract_class.abi)
-            if isinstance(contract_class, ContractClass)
+            if isinstance(contract_class, DeprecatedContractClass)
             else json.loads(cast(str, contract_class.abi))
         )
 
@@ -218,7 +218,7 @@ class ProxyResolutionError(Exception):
 
 async def _get_class_at(
     address: Address, client: Client
-) -> Union[ContractClass, SierraContractClass]:
+) -> Union[DeprecatedContractClass, SierraContractClass]:
     try:
         contract_class_hash = await client.get_class_hash_at(contract_address=address)
         contract_class = await client.get_class_by_hash(class_hash=contract_class_hash)
